refactor(quality): rename DatasetQualityScore to QualityScore and DatasetQualityItemWarning to QualityItemWarning
