+++ datagouv-components/src/components/DataserviceQuality.vue
@@ -4,7 +4,7 @@
     button-class="flex items-center gap-1 rounded-sm p-1.5 -m-1.5 border-transparent -outline-offset-2 hover:bg-gray-some transition-colors w-full text-left"
     :button-props="{ title: t('Qualité des métadonnées') }"
   >
-    <DatasetQualityScore
+    <QualityScore
       :score="quality.score"
       class="w-full"
     />
@@ -18,43 +18,43 @@
   </Toggletip>
   <template v-if="!hideWarnings">
     <ul class="list-none pl-0">
-      <DatasetQualityItemWarning
+      <QualityItemWarning
         :quality-item="quality.has_description"
         :message="t('Description non renseignée')"
       />
-      <DatasetQualityItemWarning
+      <QualityItemWarning
         :quality-item="quality.has_machine_documentation"
         :message="t('Documentation machine (OpenAPI) manquante')"
       />
-      <DatasetQualityItemWarning
+      <QualityItemWarning
         :quality-item="quality.has_technical_documentation"
         :message="t('Documentation technique manquante')"
       />
-      <DatasetQualityItemWarning
+      <QualityItemWarning
         :quality-item="quality.has_business_documentation"
         :message="t('Documentation métier manquante')"
       />
-      <DatasetQualityItemWarning
+      <QualityItemWarning
         :quality-item="quality.license"
         :message="t('Licence non renseignée')"
       />
-      <DatasetQualityItemWarning
+      <QualityItemWarning
         :quality-item="quality.has_contact_point"
         :message="t('Point de contact non renseigné')"
       />
-      <DatasetQualityItemWarning
+      <QualityItemWarning
         :quality-item="quality.has_base_api_url"
         :message="t(`URL de base de l'API non renseignée`)"
       />
-      <DatasetQualityItemWarning
+      <QualityItemWarning
         :quality-item="quality.availability_documented"
         :message="t('Disponibilité non documentée')"
       />
-      <DatasetQualityItemWarning
+      <QualityItemWarning
         :quality-item="quality.rate_limiting_documented"
         :message="t(`Limite d'appels non documentée`)"
       />
-      <DatasetQualityItemWarning
+      <QualityItemWarning
         :quality-item="quality.access_conditions_clear"
         :message="t(`Conditions d'accès non précisées`)"
       />
@@ -65,8 +65,8 @@
 <script setup lang="ts">
 import { RiInformationLine } from '@remixicon/vue'
 import type { DataserviceQuality } from '../types/dataservices'
-import DatasetQualityScore from './DatasetQualityScore.vue'
-import DatasetQualityItemWarning from './DatasetQualityItemWarning.vue'
+import QualityScore from './QualityScore.vue'
+import QualityItemWarning from './QualityItemWarning.vue'
 import DataserviceQualityTooltipContent from './DataserviceQualityTooltipContent.vue'
 import Toggletip from './Toggletip.vue'
 import { useTranslation } from '../composables/useTranslation'
