+++ datagouv-components/src/components/QualityItemWarning.vue
