+++ datagouv-components/src/components/QualityScore.vue
